refactor: extract AnnouncementBar.tsx from Header.tsx
